- authentication entities clone methods added, CloneManager removed,
- user CRUD UI refactored a bit.

+++ mgis-web/src/main/webapp/app2/admin/users/admin-users-module.js
@@ -1,12 +1,13 @@
 /**
  * Created by asd on 21/06/15.
  */
-angular.module("mgis.admin.users", [  "ui.bootstrap", "ui.multiselect", //
+angular.module("mgis.admin.users", ["ui.bootstrap", "ui.multiselect", //
+    "mgis.commons", //
     "mgis.admin.users.service", "mgis.admin.privileges.service" //
 ]) //
     .config(function ($stateProvider, $urlRouterProvider) {
     }) //
-    .controller("AdminUsersController", function ($scope, AdminUsersService, AdminPrivilegesService, $modal) {
+    .controller("AdminUsersController", function ($scope, AdminUsersService, AdminPrivilegesService, $modal, MGISCommonsModalForm, $rootScope) {
         $scope.first = 0;
         $scope.max = 15;
         function updateGrid() {
@@ -17,24 +18,14 @@ angular.module("mgis.admin.users", [  "ui.bootstrap", "ui.multiselect", //
             });
         }
 
-        function openModal() {
+        function openModal(modalScope) {
             AdminPrivilegesService.get("").then(function (availablePrivileges) {
-                $scope.availablePrivileges = availablePrivileges.list;
-                $modal.open({
-                    animate: true,
-                    scope: $scope,
-                    templateUrl: "app2/admin/users/admin-user-form.htm",
-                    controller: function ($scope, $modalInstance) {
-                        $scope.ok = function () {
-                            AdminUsersService.save($scope.editedUser).then(function () {
-                                $modalInstance.close();
-                                updateGrid();
-                            })
-                        }
-                        $scope.cancel = function () {
-                            $modalInstance.dismiss("cancel");
-                        }
-                    }
+                modalScope.availablePrivileges = availablePrivileges.list;
+                MGISCommonsModalForm.edit("app2/admin/users/admin-user-form.htm", modalScope, function ($scope, $modalInstance) {
+                    AdminUsersService.save($scope.editedUser).then(function () {
+                        $modalInstance.close();
+                        updateGrid();
+                    });
                 });
             });
         }
@@ -42,35 +33,29 @@ angular.module("mgis.admin.users", [  "ui.bootstrap", "ui.multiselect", //
         updateGrid();
 
         $scope.add = function () {
-            $scope.editedUser = {
+            var modalScope = $rootScope.$new();
+            modalScope.editedUser = {
                 id: 0,
                 username: "",
                 password: "",
                 active: true,
                 roles: []
             }
-            openModal();
+            openModal(modalScope);
         }
         $scope.edit = function (userId) {
             AdminUsersService.get(userId).then(function (user) {
-                $scope.editedUser = user;
-                openModal();
+                var modalScope = $rootScope.$new();
+                modalScope.editedUser = user;
+                openModal(modalScope);
             });
         }
         $scope.remove = function (userId) {
-            var modalInstance = $modal.open({
-                templateUrl: 'app2/common/confirm-deletion.htm',
-                controller: function ($scope, $modalInstance) {
-                    $scope.ok = function () {
-                        AdminUsersService.remove(userId).then(function () {
-                            $modalInstance.close();
-                            updateGrid();
-                        });
-                    }
-                    $scope.cancel = function () {
-                        $modalInstance.dismiss('cancel');
-                    }
-                }
+            MGISCommonsModalForm.confirmRemoval(function ($modalInstance) {
+                AdminUsersService.remove(userId).then(function () {
+                    $modalInstance.close();
+                    updateGrid();
+                });
             });
         }
     }) //
